Add files via upload
Problem-Project 1:
The new 8-puzzle has eight tiles as usual, but each tile may move Up, Up-Right, Right, Down-Right, Down, Down-Left, Left, and Up-Left. The cost of a regular move (Up, Down, Left, or Right) is 1, and the cost of a diagonal move (Up-Right, Down-Right, Down-Left, and Up-Left) is 1.4 (approximately equal to √2 ≈ 1.414).